Events

+++ LanceOfDestiny-FourthDimension/src/main/java/org/LanceOfDestiny/domain/physics/BallCollider.java
@@ -3,20 +3,20 @@ package org.LanceOfDestiny.domain.physics;
 import org.LanceOfDestiny.domain.GameObject;
 
 public class BallCollider extends Collider {
-    private float radius;
+    private double radius;
 
-    public BallCollider(Vector velocity, ColliderType colliderType, float radius, GameObject gameObject) {
+    public BallCollider(Vector velocity, ColliderType colliderType, double radius, GameObject gameObject) {
         super(velocity, colliderType, gameObject);
         this.radius = radius;
     }
 
     // Additional functionality specific to BallCollider can go here
 
-    public float getRadius() {
+    public double getRadius() {
         return radius;
     }
 
-    public void setRadius(float radius) {
+    public void setRadius(double radius) {
         this.radius = radius;
     }
 }
